Minor updates and bug fixes

+++ rtk8s/rendering/__init__.py
@@ -8,7 +8,7 @@ def render_string(s, config):
     return t.render(**config)
 
 
-def render_project(templates, target, config, markers=(".j2",)):
+def render_project(templates, target, config, markers=(".j2",), overwrite=False):
     for root, dirs, files in os.walk(templates):
         for file in files:
 
@@ -25,9 +25,12 @@ def render_project(templates, target, config, markers=(".j2",)):
                 dst = os.path.join(dst_path, filename)
 
                 with open(src, "r") as src_file:
-                    with open(dst, "w") as dst_file:
-                        s = render_string(src_file.read(), config)
-                        dst_file.write(s)
+                    if not os.path.exists(dst) or overwrite:
+                        with open(dst, "w") as dst_file:
+                            s = render_string(src_file.read(), config)
+                            dst_file.write(s)
+                    else:
+                        print(f"Skipping {dst} - it already exists!")
 
             else:
                 dst = os.path.join(dst_path, file)
